+++ src/main/webapp/components/search-forms/editor.tsx
@@ -7,12 +7,9 @@ import Typography from '@material-ui/core/Typography'
 import PlayCircleFilledIcon from '@material-ui/icons/PlayCircleFilled'
 import React, { useState } from 'react'
 import loadable from 'react-loadable'
-import {
-  AttributeDefinition,
-  QueryType,
-} from '../components/query-builder/types'
-const { useQueryExecutor, useApolloFallback } = require('../react-hooks')
-const genResults = require('../gen-results').default
+import { AttributeDefinition, QueryType } from '../query-builder/types'
+const { useQueryExecutor, useApolloFallback } = require('../../react-hooks')
+const genResults = require('../../gen-results').default
 
 const Loading = () => {
   return (
@@ -29,7 +26,7 @@ if (typeof window !== 'undefined') {
     loader: () =>
       import(//prettier-ignore
       // @ts-ignore
-      /* webpackChunkName: "visualizations" */ '../workspaces/visualizations').then(
+      /* webpackChunkName: "visualizations" */ '../../workspaces/visualizations').then(
         module => React.memo(module.default) // React.memo prevents the map from re-rendering when parent does
       ),
     loading: Loading,
