+++ src/main/webapp/components/search-forms/search-forms.stories.tsx
@@ -1,15 +1,17 @@
-const { storiesOf } = require('../@storybook/react')
+const { storiesOf } = require('../../@storybook/react')
 import React from 'react'
 import { boolean } from '@storybook/addon-knobs'
 import SearchFormRoute from './route'
 import { QueryEditor } from './editor'
 import { action } from '@storybook/addon-actions'
-import { defaultFilter } from '../components/query-builder/filter/filter-utils'
-const { SelectionProvider } = require('../react-hooks/use-selection-interface')
+import { defaultFilter } from '../query-builder/filter/filter-utils'
+const {
+  SelectionProvider,
+} = require('../../react-hooks/use-selection-interface')
 
-import QueryBuilder from '../components/query-builder/query-builder'
-import { QueryType } from '../components/query-builder/types'
-const useState = require('../@storybook/use-state').default
+import QueryBuilder from '../query-builder/query-builder'
+import { QueryType } from '../query-builder/types'
+const useState = require('../../@storybook/use-state').default
 
 const stories = storiesOf('Search Forms', module)
